Version 0.99
Bug in setting up tools whereby a tool with no heaters, or no drives, specified had
the wrong number of heaters, or drives, associated with it.  This automatically fixes
a consequent bug in the web interface, which wouldn't drive a tool with just drives
and no heaters until an M302 was sent.