Version 1.02
M563 changed so that you can redefine the heaters and drives of an existing tool with
it as well as using it to specify a new one.  Sent without any arguments it prints a list
of all the defined tools.

Bug fixed whereby the extruder drive in the web interface caused the current heater to switch off.